+++ win/rfb_win32/SInput.h
@@ -44,10 +44,10 @@ namespace rfb {
       // - Create a pointer event at a the given coordinates, with the
       //   specified button state.  The event must be specified using
       //   Screen coordinates.
-      void pointerEvent(const Point& pos, uint8_t buttonmask);
+      void pointerEvent(const Point& pos, uint16_t buttonmask);
     protected:
       Point last_position;
-      uint8_t last_buttonmask;
+      uint16_t last_buttonmask;
     };
 
     // -=- Keyboard event handling
